feat(exam):  add interface for exam service| parivash cheraghi
